FOCSproj/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
cryptography==41.0.7
pyotp==2.9.0
bcrypt==4.0.1
python-dotenv==1.0.0
scapy==2.5.0
psutil==5.9.6
Flask-Mail==0.9.1
